% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{collect_sample_altares}
\alias{collect_sample_altares}
\title{Collect sample Altares}
\usage{
collect_sample_altares(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{date}
}
\value{
a table
}
\description{
Collect sample Altares
}
\examples{
\dontrun{
collect_sample_altares(db = database_signauxfaibles, .date = "2017-01-01")
}

}
