% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{collect_sample_effectif}
\alias{collect_sample_effectif}
\title{Collect sample effectif}
\usage{
collect_sample_effectif(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{a date}
}
\value{
a table
}
\description{
Collect sample effectif
}
\examples{
\dontrun{
collect_sample_effectif(db = database_signauxfaibles, .date = "2017-01-01")
}

}
