% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute AUC}
\usage{
compute_auc(f, df_train, df_test)
}
\arguments{
\item{f}{a formula}

\item{df_train}{a training table}

\item{df_test}{a testing table}
}
\value{
a number
}
\description{
Compute AUC
}
\examples{
\dontrun{
compute_auc(f = formulas_0_12$m5, df_train = sample_train, df_test = sample_test)
}

}

