% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_filter_ccsv}
\alias{compute_filter_ccsv}
\title{Compute filter CCSV}
\usage{
compute_filter_ccsv(db, .date)
}
\arguments{
\item{db}{a database connexion}

\item{.date}{a date}
}
\value{
a table in the database
}
\description{
Compute filter CCSV
}
\examples{

\dontrun{
compute_filter_ccsv(db = database_signauxfaibles, .date = "2013-01-01")
compute_filter_ccsv(db = database_signauxfaibles, .date = "2014-01-01")
compute_filter_ccsv(db = database_signauxfaibles, .date = "2015-01-01")
}

}

