% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_filter_proccollectives}
\alias{compute_filter_proccollectives}
\title{Compute filter proccollectives}
\usage{
compute_filter_proccollectives(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{a date}
}
\value{
a table in the database
}
\description{
Compute filter proccollectives
}
\examples{
\dontrun{
compute_filter_proccollectives(db = database_signauxfaibles, .date = "2017-01-01")
}

}

