% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_prefilter_altares}
\alias{compute_prefilter_altares}
\title{Compute prefilter altares}
\usage{
compute_prefilter_altares(db, .date)
}
\arguments{
\item{.date}{}
}
\value{
a table in the database
}
\description{
Cette fonction a pour but de pouvoir retirer toutes les entreprises qui sont déjà en RJ/LJ à la date considérée.
}
\examples{
\dontrun{
compute_prefilter_altares(
db = database_signauxfaibles,
.date = "2013-01-01")
}

}
