% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample}
\alias{compute_sample}
\title{DEPRECATED Compute sample}
\usage{
compute_sample(db, .date, .fallback_date)
}
\arguments{
\item{db}{a database}

\item{.date}{a date}

\item{.fallback_date}{the fallback date for the table effectif}
}
\value{
a table in the database
}
\description{
DEPRECATED Compute sample
}
\examples{

\dontrun{
compute_sample(db = database_signauxfaibles, .date = "2017-01-01")
}

}
