% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_altares}
\alias{compute_sample_altares}
\title{Compute sample ALTARES}
\usage{
compute_sample_altares(db, .date)
}
\arguments{
\item{db}{name of the database}

\item{.date}{date}
}
\value{
a table in the database
}
\description{
Compute sample ALTARES
}
\examples{
\dontrun{
compute_sample_altares(
db = database_signauxfaibles,
.date = "2013-01-01"
)
}

}

