% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_apart}
\alias{compute_sample_apart}
\title{Compute sample apart}
\usage{
compute_sample_apart(db, .date, n_months = 12)
}
\arguments{
\item{db}{name of the database}

\item{.date}{initial date}

\item{n_months}{number of months to consider}
}
\value{
a table in the database
}
\description{
Compute sample apart
}
\examples{
\dontrun{
compute_sample_apart(
db = database_signauxfaibles,
initial_date = "2013-01-01",
n_months = 12
)
}

}

