% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_apart_consommee}
\alias{compute_sample_apart_consommee}
\title{Compute sample activite partielle consommee}
\usage{
compute_sample_apart_consommee(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{date}
}
\value{
a table in the databaase
}
\description{
Compute sample activite partielle consommee
}
\examples{

\dontrun{
compute_sample_apart_consommee(db = database_signauxfaibles, .date = "2017-03-01")
}

}
