% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_delais}
\alias{compute_sample_delais}
\title{Compute sample delais}
\usage{
compute_sample_delais(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{date}
}
\value{
a table in the database
}
\description{
Compute sample delais
}
\examples{
\dontrun{
compute_sample_delais(db = database_signauxfaibles, .date = "2017-01-01")
}

}

