% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_dettecumulee}
\alias{compute_sample_dettecumulee}
\title{Compute sample dette cumulee}
\usage{
compute_sample_dettecumulee(db, .date)
}
\arguments{
\item{db}{name of the database}

\item{.date}{a date}
}
\value{
a table in the database
}
\description{
Compute sample dette cumulee
}
\examples{
\dontrun{
compute_sample_dettecumulee(db = database_signauxfaibles, .date = "2013-01-01")
}

}
