% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_effectif}
\alias{compute_sample_effectif}
\title{Compute sample effectif}
\usage{
compute_sample_effectif(db, .date, .periode)
}
\arguments{
\item{db}{the name of the database}

\item{.date}{a date in the form "yyyy-mm-dd"}
}
\value{
a table in the database
}
\description{
Cette fonction permet de créer une table avec les effectifs d'un établissement à une date donnée.
}
\examples{

\dontrun{
compute_sample_effectif(
db = database_signauxfaibles,
.date = "2013-01-01")
}

}

