% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_growth_dettecumulee}
\alias{compute_sample_growth_dettecumulee}
\title{Compute sample growth dette cumulee}
\usage{
compute_sample_growth_dettecumulee(db, .date, lag, name)
}
\arguments{
\item{db}{name of the database}

\item{.date}{date}

\item{lag}{number of lag months}
}
\value{
a table in the database
}
\description{
Compute sample growth dette cumulee
}
\examples{
\dontrun{
compute_sample_growth_dettecumulee(
db = database_signauxfaibles,
date = "2013-01-01",
lag = 12,
)
}
}
