% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_lag_dettecumulee}
\alias{compute_sample_lag_dettecumulee}
\title{Compute sample lag dette cumulee}
\usage{
compute_sample_lag_dettecumulee(db, .date, lag)
}
\arguments{
\item{db}{database}

\item{.date}{a date}

\item{lag}{number of months}
}
\value{
a table in the database
}
\description{
Compute sample lag dette cumulee
}
\examples{

\dontrun{
compute_sample_lag_dettecumulee(db = database_signauxfaibles, lag = 12, .date = "2017-01-01")
}

}

