% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_meancotisation}
\alias{compute_sample_meancotisation}
\title{Compute sample cotisations}
\usage{
compute_sample_meancotisation(db, .date)
}
\arguments{
\item{db}{name of the database}

\item{.date}{a date to be considered}
}
\value{
a table in the database
}
\description{
Compute sample cotisations
}
\examples{
\dontrun{
compute_sample_meancotisation(db = database_signauxfaibles, .date = "2017-01-01")
}

}

