% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_nbdebits}
\alias{compute_sample_nbdebits}
\title{Compute sample nbdebits}
\usage{
compute_sample_nbdebits(db, .date, n_months)
}
\arguments{
\item{db}{database}

\item{.date}{date}

\item{n_months}{n_months}
}
\value{
a table in the database
}
\description{
Compute sample nbdebits
}
\examples{

\dontrun{
compute_sample_nbdebits(
db = database_signauxfaibles,
.date = "2017-01-01",
n_months = 12)
}

}

