% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_sample_sirene}
\alias{compute_sample_sirene}
\title{Compute sample sirene}
\usage{
compute_sample_sirene(db)
}
\arguments{
\item{db}{name of the database}
}
\value{
a table in the database
}
\description{
Compute sample sirene
}
\examples{
\dontrun{
compute_sample_sirene(
db = database_signauxfaibles,
)
}

}

