% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample}
\alias{compute_wholesample}
\title{Compute whole sample}
\usage{
compute_wholesample(db, name)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}
}
\value{
a table in the database
}
\description{
Compute whole sample
}
\examples{
\dontrun{
compute_wholesample(
db = database_signauxfaibles,
name = "wholesample")
}

}
