% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_apart}
\alias{compute_wholesample_apart}
\title{Compute wholesample Activité partielle}
\usage{
compute_wholesample_apart(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the database}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample Activité partielle
}
\examples{

\dontrun{
compute_wholesample_apart(
db = database_signauxfaibles,
name = "wholesample_apart",
start = "2013-01-01",
end = "2017-03-01")
}

}

