% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_apartconsommee}
\alias{compute_wholesample_apartconsommee}
\title{Compute wholesample activite partielle consommmee}
\usage{
compute_wholesample_apartconsommee(db, name, start, end)
}
\arguments{
\item{db}{name of the database}

\item{name}{name of the table}

\item{start}{starting date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample activite partielle consommmee
}
\examples{

\dontrun{
compute_wholesample_apartconsommee(
db = database_signauxfaibles,
name = "wholesample_apartconsommee",
start = "2013-01-01",
end = "2017-03-01"
)
}

}

