% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_banquedefrance}
\alias{compute_wholesample_banquedefrance}
\title{Compute wholesample banque de France}
\usage{
compute_wholesample_banquedefrance(db, name)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}
}
\value{
a table
}
\description{
Compute wholesample banque de France
}
\examples{

\dontrun{
compute_wholesample_banquedefrance(db = database_signauxfaibles, name = "table_banquedefrance")
}

}
