% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_ccsv}
\alias{compute_wholesample_ccsv}
\title{Compute wholesample CCSV}
\usage{
compute_wholesample_ccsv(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample CCSV
}
\examples{

compute_wholesample_ccsv(
db = database_signauxfaibles,
name = "wholesample_ccsv",
start = "2013-01-01",
end = "2017-03-01"
)

}
