% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_delais}
\alias{compute_wholesample_delais}
\title{Compute wholesample delais}
\usage{
compute_wholesample_delais(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample delais
}
\examples{

\dontrun{
compute_wholesample_delais(
db = database_signauxfaibles,
name = "wholesample_delais",
start = "2013-01-01",
end = "2017-03-01"
)
}

}

