% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_dettecumulee}
\alias{compute_wholesample_dettecumulee}
\title{Compute wholesample dettecumulee}
\usage{
compute_wholesample_dettecumulee(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the output table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample dettecumulee
}
\examples{

\dontrun{
compute_wholesample_dettecumulee(
db = database_signauxfaibles,
name = "wholesample_dettecumulee",
start = "2013-01-01",
end = "2017-03-01"
)
}

}
