% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_dettecumulee_12m}
\alias{compute_wholesample_dettecumulee_12m}
\title{compute wholesample dettecumulee 12m}
\usage{
compute_wholesample_dettecumulee_12m(db, name, start, end)
}
\arguments{
\item{db}{a database conexion}

\item{name}{name of the table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
compute wholesample dettecumulee 12m
}
\examples{

}

