% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_effectif}
\alias{compute_wholesample_effectif}
\title{Compute whole sample effectif}
\usage{
compute_wholesample_effectif(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}
}
\value{
a table in the database
}
\description{
Compute whole sample effectif
}
\examples{

\dontrun{
compute_wholesample_effectif(db = database_signauxfaibles, name = "wholesample_effectif")
}

}
