% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_lagdettecumulee}
\alias{compute_wholesample_lagdettecumulee}
\title{Compute wholesample lag dette cumulee}
\usage{
compute_wholesample_lagdettecumulee(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compute wholesample lag dette cumulee
}
\examples{

\dontrun{
compute_wholesample_lagdettecumulee(
db = database_signauxfaibles,
name = "wholesample_lagdettecumulee",
start = "2013-01-01",
end = "2017-03-01")
}

}

