% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_meancotisation}
\alias{compute_wholesample_meancotisation}
\title{Compute wholesample meancotisation}
\usage{
compute_wholesample_meancotisation(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the output table}

\item{start}{start date}

\item{end}{end date}
}
\value{
table in the database
}
\description{
Compute wholesample meancotisation
}
\examples{
\dontrun{
compute_wholesample_meancotisation(
db = database_signauxfaibles,
name = "wholesample_meancotisation",
start = "2013-01-01",
end = "2017-03-01"
)
}

}

