% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_urssaf_date}
\alias{convert_urssaf_date}
\title{convert_urssaf_date}
\usage{
convert_urssaf_date(weird_date)
}
\arguments{
\item{weird_date}{}
}
\value{
a date
}
\description{
convert_urssaf_date
}
\examples{
convert_urssaf_date(weird_date = "1010115")

}
