% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urssaf.R
\name{convert_urssaf_periods_}
\alias{convert_urssaf_periods_}
\title{Convert urssaf periods to standard format}
\usage{
convert_urssaf_periods_(.data, .variable, format = "yyqm")
}
\arguments{
\item{.data}{a table with a period}

\item{.variable}{the name of the variable with urssaf periods}

\item{format}{can bee yyqm or yyyyqm}
}
\value{
an new table with two more columns : periodicity and period
}
\description{
Urssaf store dates in a very specific format.
This can be years + 62 for annual values,
years + quarter + 0 for quarterly values
years + quarter + number of the month in the quarter for monthly values.
}
\examples{

\dontrun{
convert_urssaf_periods_(.variable = ~ yyyyqm, format = "yyyyqm")
}

}

