% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgre.R
\name{database_connect}
\alias{database_connect}
\title{Database connect}
\usage{
database_connect(file = "keys.json")
}
\arguments{
\item{file}{name of the file where keys are stored}
}
\value{
a connection to the signauxfaible database
}
\description{
This function reads the file keys.json at the root of the directory and create a connection to the postgre database.
}
\details{
The file keys.json should have the following format :

{
"host": ["127.0.0.1"],
"dbname": ["databasename"],
"port": ["5433"],
"id":["login"],
"pw":["password"]
}
}
\examples{
database_signauxfaibles <- database_connect()

}

