% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{db_drop_table_ifexist}
\alias{db_drop_table_ifexist}
\title{Drop table if exist}
\usage{
db_drop_table_ifexist(db, table)
}
\arguments{
\item{db}{name of the database}

\item{table}{name of the table}
}
\description{
The default function db_drop_table in the dplyr package returns an error if the table already exist. This creates a lot of errors in the data manipulation pipe. db_drop_table_ifexist check if the table exist and drop the table if and only if it exist.
}
\examples{
db_drop_table_ifexist(
db = database_signauxfaibles,
table = "table_periods"
)

}
