% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_infinite}
\alias{detect_infinite}
\title{Detect infinite values}
\usage{
detect_infinite(table)
}
\arguments{
\item{table}{a table}
}
\value{
a tibble
}
\description{
Detect infinite values
}
\examples{

}
