% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_na}
\alias{detect_na}
\title{Detect missing values}
\usage{
detect_na(table)
}
\arguments{
\item{table}{name of the input variable}
}
\value{
a table with 3 columns (variable, n_missing, share_missing)
}
\description{
Detect na takes a table as input and returns a table with the number and the share of missing values by column
}
\examples{

tibble::tibble(col1 = c(1,NA,3,4), col2 = c(NA,NA,NA,NA)) \%>\%
detect_na()

}

