% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_accountnumber}
\alias{get_accountnumber}
\title{Get accountnumber}
\usage{
get_accountnumber(db, .siret)
}
\arguments{
\item{db}{a database connexion}

\item{.siret}{a siret number}
}
\value{
a string
}
\description{
Get accountnumber
}
\examples{

\dontrun{
get_accountnumber(
db = database_signauxfaibles,
.siret = "33289883200040"
)
}

}
