% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_ccsv}
\alias{get_ccsv}
\title{Get CCSV}
\usage{
get_ccsv(db, .siret)
}
\arguments{
\item{db}{name of the database}

\item{.siret}{a siret number}
}
\value{
a tibble
}
\description{
Get information in the CCSV table for a siret number
}
\examples{

}

