% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_delais}
\alias{get_delais}
\title{Get delais}
\usage{
get_delais(db, siret)
}
\arguments{
\item{db}{database connexion}

\item{siret}{a valid siret number}
}
\value{
a table
}
\description{
Get delais
}
\examples{

\dontrun{
get_delais(db = database_signauxfaibles, siret = "54200362900020")
}
}

