% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_raisonsociale}
\alias{get_raisonsociale}
\title{Get raison sociale}
\usage{
get_raisonsociale(db, .siret)
}
\arguments{
\item{db}{a database connexion}

\item{.siret}{a siret number}
}
\value{
a string
}
\description{
Get raison sociale
}
\examples{


}

