% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_ratio_cotisation_effectif}
\alias{get_ratio_cotisation_effectif}
\title{Get ratio cotisation effectif}
\usage{
get_ratio_cotisation_effectif(db, siret)
}
\arguments{
\item{db}{database connexion}

\item{siret}{siret number}
}
\value{
a table
}
\description{
Get ratio cotisation effectif
}
\examples{

}

