% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_ratio_dettecumulee_cotisation}
\alias{get_ratio_dettecumulee_cotisation}
\title{Get ratio dettecumulee cotisation}
\usage{
get_ratio_dettecumulee_cotisation(db, siret)
}
\arguments{
\item{db}{a database connexion}

\item{siret}{a siret number}
}
\value{
a table
}
\description{
Get ratio dettecumulee cotisation
}
\examples{

}

