% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_sirene}
\alias{get_sirene}
\title{Get SIRENE}
\usage{
get_sirene(db, .siret)
}
\arguments{
\item{db}{name of the database}

\item{.siret}{a siret number}
}
\value{
a list
}
\description{
get a list of information from the SIRENE database
}
\examples{
\dontrun{
get_sirene(db = database_signauxfaibles, .siret = '40094678600011')
}

}
