% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_siret}
\alias{get_siret}
\title{Get siret}
\usage{
get_siret(db, .numero_compte)
}
\arguments{
\item{db}{database connexion}

\item{.numero_compte}{account number}
}
\value{
a string
}
\description{
Get siret
}
\examples{

\dontrun{
get_siret(
db = database_signauxfaibles,
.numero_compte = "267000001600093120"
)
}

}
