% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_table_last_n_months}
\alias{get_table_last_n_months}
\title{Get table last n months}
\usage{
get_table_last_n_months(.date, .n_months)
}
\arguments{
\item{.date}{a date as a string}

\item{.n_months}{an integer}
}
\value{
a tibble
}
\description{
Get table last n months
}
\examples{
get_table_last_n_months(.date = "2017-01-01", .n_months = 12)

}
