% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgre.R
\name{has_variable}
\alias{has_variable}
\title{Has variable}
\usage{
has_variable(db, table, variable)
}
\arguments{
\item{db}{database connexion}

\item{table}{table in the database}

\item{variable}{variable to be checked}
}
\value{
a boolean
}
\description{
Has variable
}
\examples{
\dontrun{
has_variable(db = database_signauxfaibles, table = "table_altares", variable = "siret")
}
}
