% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_apart_heuresconsommees}
\alias{import_apart_heuresconsommees}
\title{Import Apart heures consommées}
\usage{
import_apart_heuresconsommees(path)
}
\arguments{
\item{path}{path to the file}
}
\value{
a tibble
}
\description{
Import Apart heures consommées
}
\examples{

\dontrun{
import_apart_heuresconsommees(path = "raw-data/direccte/act_partielle_consommée.xlsx")
}

}
