% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_activite_partielle}
\alias{import_table_activite_partielle}
\title{Import table activite partielle}
\usage{
import_table_activite_partielle(path)
}
\arguments{
\item{path}{path}
}
\value{
a tibble
}
\description{
Import de la table d'activité partielle augmentée
}
\examples{

\dontrun{
import_table_activite_partielle(path = "raw-data/direccte/act_partielle_ddes_2012_mai2017.xlsx")
}

}
