% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_altares}
\alias{import_table_altares}
\title{Import table Altares}
\usage{
import_table_altares(path)
}
\arguments{
\item{path}{the path to the data}
}
\value{
a tibble with all variables in the table altares
}
\description{
Import table Altares
}
\examples{
\dontrun{
table_altares <- import_table_altares(path = "raw-data/direccte/RECAP_ALTARES.csv")
}

}
