% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_apart}
\alias{import_table_apart}
\title{Import table apart  (ie Activite partielle)}
\usage{
import_table_apart(path)
}
\arguments{
\item{path}{path to the excel file}
}
\value{
a tibble
}
\description{
Import table apart  (ie Activite partielle)
}
\examples{
\dontrun{
import_table_apart(path = "raw-data/direccte/201609_POP_Bourg Fr Comté de janv2008àaoût2016.xls")
}

}

