% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_ccsv}
\alias{import_table_ccsv}
\title{Import table CCSV}
\usage{
import_table_ccsv(path)
}
\arguments{
\item{path}{chemin vers le fichier CSV}
}
\value{
a tibble : table de données
}
\description{
Cette fonction permet d'importer les données CCSV dans la base de données.
}
\examples{

\dontrun{
"raw-data/urssaf/Bourgogne_ccsf.csv" \%>\%
import_table_ccsv(path = .)
}

\dontrun{
"raw-data/urssaf/FRC_ccsf.csv" \%>\%
import_table_ccsv(path = .)
}

}
