% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_cotisation_csv}
\alias{import_table_cotisation_csv}
\title{Import table cotisation in a CSV format}
\usage{
import_table_cotisation_csv(path)
}
\arguments{
\item{path}{path to the CSV file}
}
\value{
a tibble
}
\description{
Import table cotisation in a CSV format
}
\examples{

\dontrun{
import_table_cotisation_csv(
path = "raw-data/urssaf/Urssaf_bourgogne_Cotis_dues_09_2016_01_2017.csv"
)
}

}

