% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_debit}
\alias{import_table_debit}
\title{Import table debits}
\usage{
import_table_debit(path)
}
\arguments{
\item{path}{the path to the debit files}
}
\value{
a tibble
}
\description{
Import table debits
}
\examples{
\dontrun{
import_table_debit(path = "raw-data/urssaf/Debit_bourgogne_31_07_2016.csv")
}

}

