% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_delais}
\alias{import_table_delais}
\title{Import table délais}
\usage{
import_table_delais(path)
}
\arguments{
\item{path}{path to the sources files}
}
\value{
a tibble
}
\description{
Import table délais
}
\examples{

\dontrun{
import_table_delais(path = "raw-data/urssaf/delais_bourgogne_01_2013_01_2017_utf8.csv")
import_table_delais(path = "raw-data/urssaf/delais_franchecomte_01_2013__01_2017.csv")
}

}

