% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_effectif}
\alias{import_table_effectif}
\title{Import table effectif}
\usage{
import_table_effectif(path)
}
\arguments{
\item{path}{path of the source files. Source files are stored in raw-data/urssaf directory}
}
\value{
a tibble with columns siret, raison_sociale, period, effectif, code_departement
}
\description{
Import table effectif
}
\examples{
\dontrun{
import_table_effectif(
path = "raw-data/urssaf/Urssaf_emploi_BFC_200401_201606.csv"
)
}

}

