% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_naf}
\alias{import_table_naf}
\title{Import table naf}
\usage{
import_table_naf(path)
}
\value{
a tibble with code nav level 5, code naf level 1 and label naf level 1
}
\description{
This function imports the NAF code http://www.insee.fr/fr/methodes/default.asp?page=nomenclatures/naf2008/naf2008.htm
}
\examples{
\dontrun{
import_table_naf(path = "data-raw/insee/naf/naf2008_5_niveaux.xls")
}

}
