% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgre.R
\name{insert_multi}
\alias{insert_multi}
\title{Insert a dataframe given a number of slices}
\usage{
insert_multi(dest, name, df, slices, ...)
}
\arguments{
\item{df}{dataframe to insert}

\item{slices}{number of slices}

\item{db}{name of the database}

\item{table}{name of the table}
}
\description{
Alone, copy_to kills memory.
}
