% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{is_ccsf}
\alias{is_ccsf}
\title{is ccsf}
\usage{
is_ccsf(db, siret)
}
\arguments{
\item{db}{a database connexion}

\item{siret}{a valid siret number}
}
\value{
a boolean
}
\description{
is ccsf
}
\examples{

}
