% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_fake_sequence}
\alias{make_fake_sequence}
\title{Make fake sequence}
\usage{
make_fake_sequence(start, end, last)
}
\arguments{
\item{start}{a start date}

\item{end}{an end date}

\item{last}{the last update}
}
\value{
a character vector
}
\description{
Cette fonction permet de générer une fausse séquence.
C'est utile pour générer l'échantillon sur les effectifs
lorsque la date de dernière mise à jour est en retard
par rapport aux autres datasets.
}
\examples{

make_fake_sequence(start = "2013-01-01", end = "2017-03-01", last = "2017-01-01")

}
