% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_sequence}
\alias{make_sequence}
\title{Make sequence}
\usage{
make_sequence(start, end)
}
\arguments{
\item{start}{start date}

\item{end}{end date}
}
\value{
a vector of dates as a string
}
\description{
Make sequence
}
\examples{

make_sequence(start = "2013-01-01", end = "2017-03-01")

}

