% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgre.R
\name{map_variables}
\alias{map_variables}
\title{Map variables}
\usage{
map_variables(db)
}
\arguments{
\item{db}{database connexion}
}
\value{
a table
}
\description{
Map variables
}
\examples{
\dontrun{
map_variables(db = database_signauxfaibles)
}
}

