% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{plot_effectif}
\alias{plot_effectif}
\title{Plot effectif}
\usage{
plot_effectif(db, .siret)
}
\arguments{
\item{db}{a database connexion}

\item{.siret}{siret number}
}
\value{
a ggplot
}
\description{
Plot effectif
}
\examples{
\dontrun{
plot_effectif(db = database_signauxfaibles, .siret = "40094678600011")
}

}
