% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{plot_meancotisation}
\alias{plot_meancotisation}
\title{Plot mean cotisation}
\usage{
plot_meancotisation(db, siret)
}
\arguments{
\item{db}{database connexion}

\item{siret}{a siret number}
}
\value{
a ggplot
}
\description{
Plot mean cotisation
}
\examples{

}

