% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{plot_ratio_cotisation_effectif}
\alias{plot_ratio_cotisation_effectif}
\title{PLot ratio cotisation effectif}
\usage{
plot_ratio_cotisation_effectif(db, siret)
}
\arguments{
\item{db}{a database connexion}

\item{siret}{a siret number}
}
\value{
a ggplot
}
\description{
PLot ratio cotisation effectif
}
\examples{

}

