% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{plot_ratio_dettecumulee_cotisation}
\alias{plot_ratio_dettecumulee_cotisation}
\title{plot ratio dettecumulee cotisation}
\usage{
plot_ratio_dettecumulee_cotisation(db, siret)
}
\arguments{
\item{db}{database connexion}

\item{siret}{siret number}
}
\value{
a ggplot
}
\description{
plot ratio dettecumulee cotisation
}
\examples{

}

