% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot roc curve}
\usage{
plot_roc_curve(model, table)
}
\arguments{
\item{model}{the name of a model}

\item{table}{the name of the test data}
}
\value{
a ggplot graph
}
\description{
Plot roc curve
}
\examples{
\dontrun{
plot_roc_curve(model = model_departements, table = table_test_preprocessed)
}

}
