% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_lastperiod}
\alias{return_lastperiod}
\title{Return lastperiods}
\usage{
return_lastperiod(x, sequence1, sequence2)
}
\arguments{
\item{x}{a date}

\item{sequence1}{a sequence}

\item{sequence2}{a sequence}
}
\value{
a date
}
\description{
Return lastperiods
}
\examples{

return_lastperiod(
x = "2017-02-01",
sequence1 = make_sequence(start = "2013-01-01", end = "2017-03-01"),
sequence2 = make_fake_sequence(start = "2013-01-01", end = "2017-03-01", last = "2017-01-01")
)

}

