% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{area_under_curve}
\alias{area_under_curve}
\title{Area under curve}
\usage{
area_under_curve(model, table)
}
\arguments{
\item{model}{a model}

\item{table_test}{the name of the table including test data}
}
\value{
a scalar value with area under curve
}
\description{
Area under curve
}
\examples{
\dontrun{
area_under_curve(model = model_1, table_test = table_test_preprocessed)
}


}
