% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_prefilter_altares}
\alias{compute_prefilter_altares}
\title{Compute prefilter Altares}
\usage{
compute_prefilter_altares(db, .date)
}
\arguments{
\item{db}{database}

\item{.date}{date}
}
\value{
a table in the database
}
\description{
Compute prefilter Altares
}
\examples{
\dontrun{
compute_prefilter_altares(
db = database_signauxfaibles,
.date = "2017-01-01")
}

}

