% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urssaf.R
\name{convert_urssaf_date}
\alias{convert_urssaf_date}
\title{Convert URSSAF weird dates to normal dates}
\usage{
convert_urssaf_date(weird_date)
}
\arguments{
\item{weird_date}{a date in the URSSAF format}
}
\value{
a date
}
\description{
Convert URSSAF weird dates to normal dates
}
\examples{

convert_urssaf_date(weird_date = "1010115")

}

