% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_effectif}
\alias{get_effectif}
\title{Get effecitf}
\usage{
get_effectif(db, .siret)
}
\arguments{
\item{db}{a database connexion}

\item{.siret}{a siret number}
}
\value{
a tibble
}
\description{
Get the effectif table for a siret number
}
\examples{
\dontrun{
get_effectif(db = database_signauxfaibles, .siret = "40094678600011")
}

}
