% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_meancotisation}
\alias{get_meancotisation}
\title{Get meancotisation}
\usage{
get_meancotisation(db, siret)
}
\arguments{
\item{db}{database connexion}

\item{siret}{a valid siret number}
}
\description{
Get meancotisation
}
\examples{

}
