% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{\%m+\%}
\alias{\%m-\%}
\alias{reexports}
\alias{sql}
\alias{uq}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{sql}}}

  \item{lazyeval}{\code{\link[lazyeval]{uq}}}

  \item{lubridate}{\code{\link[lubridate]{\%m+\%}}, \code{\link[lubridate]{\%m-\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\keyword{internal}

